refactor: split engine into focused renderers and improve cartesian layering
